+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - remove the unnecessary type and add the html-safe tag at the rendering site
+- dummy/app/controllers/besluit-sample.ts
 
 ### Removed
 - remove ensure-changelog github action